# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product API",
        "description": "API для управления товарами, арендой и покупками в системе Laravel Product",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Создает нового пользователя в системе",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Данные для регистрации пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации или создания пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "description": "Выполняет вход пользователя в систему и возвращает токен доступа",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Данные для входа в систему",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход из системы",
                "description": "Выполняет выход пользователя из системы и аннулирует токен",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Успешный выход из системы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "title": "Ответ с ошибкой",
                "description": "Стандартный ответ API при ошибке",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Товар не найден"
                    },
                    "errors": {
                        "description": "Детали ошибок (опционально)",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Ответ с ошибками валидации",
                "description": "Ответ API при ошибках валидации",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Ошибка валидации данных"
                    },
                    "errors": {
                        "description": "Детали ошибок валидации",
                        "properties": {
                            "name": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Имя обязательно"
                                ]
                            },
                            "purchase_price": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Цена покупки должна быть числом"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "Успешный ответ",
                "description": "Стандартный успешный ответ API",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Операция выполнена успешно"
                    },
                    "data": {
                        "description": "Данные ответа (если необходимо)",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Запрос регистрации",
                "description": "Данные для регистрации нового пользователя",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "description": "Email пользователя",
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "password": {
                        "description": "Пароль пользователя",
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Запрос авторизации",
                "description": "Данные для входа в систему",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email пользователя",
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "password": {
                        "description": "Пароль пользователя",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "Пользователь",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор пользователя",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "description": "Email пользователя",
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "title": "Ответ аутентификации",
                "description": "Ответ при успешной аутентификации с токеном и данными пользователя",
                "properties": {
                    "success": {
                        "description": "Статус успешности операции",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Login successful"
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "token": {
                                "description": "Токен доступа",
                                "type": "string",
                                "example": "1|abcdef123456789..."
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Аутентификация через Laravel Sanctum токены",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}